+++ nym-vpn-android/billing/src/main/java/net/nymtech/billing/model/BillingInfo.kt
@@ -1,0 +1,7 @@
+package net.nymtech.billing.model
+
+interface BillingInfo {
+	val responseCode: BillingCode
+	val debugMessage: String
+		get() = ""
+}