+++ nym-vpn-android/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - In "Gateway details", use relative time in last update. (https://github.com/nymtech/nym-vpn-client/pull/3748)
 - Display server up-time. (https://github.com/nymtech/nym-vpn-client/pull/3749)
 - Add streaming icon on home and Exit + update tooltip (https://github.com/nymtech/nym-vpn-client/pull/3774)
+- Added In App purchases (https://github.com/nymtech/nym-vpn-client/pull/3754)
+- Passphrase screen added (https://github.com/nymtech/nym-vpn-client/pull/3754)
+- UI updates and changes (https://github.com/nymtech/nym-vpn-client/pull/3754)
+- Account creation flow added (https://github.com/nymtech/nym-vpn-client/pull/3754)
 
 ## [2.1.0]